iot-sensors/requirements.txt
flask_restplus==0.13.0
pika==1.2.0
Werkzeug==2.0.1
PyYAML==5.4.1
